Add basic filesystem, Implement DASIC (Jasic), misc
The files located in the /filesystem folder next to a built jar (or on
the desktop in the /hackertyperneoneo/filesystem folder if not built)
will be used as the game's accessible filesystem.
Added commands to manipulate files/folders.
Included /defaultFilesystem in the repo
Implemented run command that runs DASIC (upgraded Jasic) files. See
https://github.com/PieKing1215/HackerTyperNeoNeo/wiki/DASIC.
Added glitch effects.
Added boot sequence.
Optimizations.